chore: validating presense of requried env vars
