+++ test/consts/FromEnvVars.ts
@@ -1,0 +1,4 @@
+import "dotenv/config"
+
+export const SERVER_HOST = process.env.SERVER_HOST || '127.0.0.1'
+export const SERVER_PORT = process.env.SERVER_PORT || 3000
